Fixed security handling to force security detection in settings activity. integrated into boothandler and timeout receiver, also set up a logic that calls disable on the non-secure one, then calls re-enable 50 ms later to ensure only the secure on is seen by user...
This now uses a SECURITY pref that is flagged on when services aren't running. Now we don't have to rely on changing the pref every time we suppress pattern requirement.

remaining todo is remove the explanation part of the How to text as it is embedded into the checkbox now.